fix: :art: Fix Wide Screen Issue
The current width of the menu will always take up 80% of the body element's width. On a very wide screen, the coffee and dessert appear far apart from their prices.  Add a max-width property to the menu class with a value of 500px to prevent it from growing too wide.